requirements.txt
paho-mqtt>=1.5.0
elasticsearch==7.10.0
